+++ Assets/Clues/ClueJournalManager.cs
@@ -29,6 +29,7 @@ namespace EscapeRoom
             }
 
             Instance = this;
+            DontDestroyOnLoad(gameObject);
         }
 
         public bool AddClue(ClueData clueData)